kong-konga-keycloak/Users/requirements.txt
pyjwt[crypto]
requests
python-keycloak
python-multipart
python-consul==1.1.0